Add JSON Text instance for (de)serializing Text strings
Add a Data.Text JSON instance to make it possible to serialize and
deserialize data types containing Data.Text strings.

Currently only Strings are supported which makes it cumbersome to use
Data.Text strings together with the json library.

This change adds a dependency to the 'text' library which is nowadays
considered stable & the de facto way to work with unicode strings.