Do not escape non-ASCII characters.
The JSON format supports unicode characters, and the pretty
printer should print the strings as they are---this is especially
useful when working with text in non-latin alphabets.

Since nowadays we have fairly standard ways to transmit
non-ASCII text  (e.g. UTF-8), there's no good reason to try
to stick to ASCII.